grpc-demo/requirements.txt
grpcio==1.6.0
grpcio-tools==1.6.0
protobuf==3.4.0
six==1.10.0
